refactor: PromptBuilder를 application에서 infrastructure로 이동
- TextRefinePromptBuilder/ImagePromptBuilder/MusicPromptBuilder를 각 모듈의 infrastructure 패키지로 이동
- 벤더별 프롬프트 포맷팅은 application이 알 필요 없는 어댑터 디테일이라 infra 내부 협력 객체로 재배치
